web-user/backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
requests>=2.31.0
pyyaml>=6.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
ruff>=0.1.0
flake8>=6.0.0
anthropic
python-dotenv
